Extract Pane and TerminalExecutionTracker out of SplitNode.swift
Pure movement, staged as PR 1 of the foreground-capture refactor:
SplitNode.swift had grown to 1564 lines holding three unrelated things —
the split tree, the ~300-line execution-tracking state machine, and Pane
with all its naming state. Each type now lives in its own file (matching
the one-type-per-test-file convention), verified byte-identical against
the original line ranges.

The one non-movement change: the probe-request plumbing scattered across
Pane (remoteProbePending + remoteProbeRetryBudget with per-field rules)
is now a RemoteProbeRequest value type owning the request/consume/retry
mechanics, with Pane keeping its public API as thin forwards (and the
isRemote / never-named guards) so no caller or test changes.
